requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
httpx==0.27.0
pydantic==2.8.2
pydantic-settings==2.3.4
SQLAlchemy==2.0.31
psycopg[binary]==3.2.1
alembic==1.13.2
python-dateutil==2.9.0.post0
loguru==0.7.2
tenacity==8.3.0
typer==0.12.3
pytest==8.2.2
pytest-asyncio==0.23.8
pytest-cov==5.0.0
pre-commit==3.7.1
ruff==0.5.5
httpx>=0.24.1